refactor: Move registerClientPlugins to separate file
cozy-helper.js includes many other files by ricochet, we don't want them in the build output
