+++ src/drive/lib/registerClientPlugins.js
@@ -1,0 +1,7 @@
+import { RealtimePlugin } from 'cozy-realtime'
+
+const registerClientPlugins = client => {
+  client.registerPlugin(RealtimePlugin)
+}
+
+export default registerClientPlugins
