+++ src/drive/targets/mobile/InitAppMobile.spec.jsx
@@ -7,10 +7,10 @@ import Alerter from 'cozy-ui/transpiled/react/Alerter'
 jest.mock('drive/mobile/lib/cozy-helper', () => {
   return {
     getLang: () => 'en',
-    initClient: jest.fn(),
-    registerClientPlugins: jest.fn()
+    initClient: jest.fn()
   }
 })
+jest.mock('drive/lib/registerClientPlugins', () => jest.fn())
 jest.mock('drive/store/configureStore')
 jest.mock('drive/lib/reporter')
 jest.mock(
